+++ ninja-core/src/test/java/ninja/session/SessionImplTest.java
@@ -32,6 +32,8 @@ import ninja.utils.NinjaConstant;
 import ninja.utils.NinjaProperties;
 import ninja.utils.SecretGenerator;
 import org.hamcrest.CoreMatchers;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.CoreMatchers.not;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -190,7 +192,6 @@ public class SessionImplTest {
 
         // verify some stuff on the set cookie
         assertEquals(true, cookieCaptor.getValue().isSecure());
-
     }
 
     @Test
@@ -442,6 +443,85 @@ public class SessionImplTest {
 
         assertNull(sessionCookie3.get(Session.EXPIRY_TIME_KEY));
     }
+    
+    @Test
+    public void testThatCookieDoesNotUseApplicationDomainWhenNotSet() {
+        when(ninjaProperties.get(NinjaConstant.applicationCookieDomain)).thenReturn(null);
+        Session sessionCookie = createNewSession();
+        sessionCookie.init(context);
+        sessionCookie.put("anykey", "anyvalue");
+
+        sessionCookie.save(context);
+
+        verify(context).addCookie(cookieCaptor.capture());
+        Cookie cookie = cookieCaptor.getValue();
+        assertThat(cookie.getDomain(), CoreMatchers.equalTo(null));
+    }
+    
+    @Test
+    public void testThatCookieUseApplicationDomain() {
+        when(ninjaProperties.get(NinjaConstant.applicationCookieDomain)).thenReturn("domain.com");
+        Session sessionCookie = createNewSession();
+        sessionCookie.init(context);
+        sessionCookie.put("anykey", "anyvalue");
+
+        sessionCookie.save(context);
+
+        verify(context).addCookie(cookieCaptor.capture());
+        Cookie cookie = cookieCaptor.getValue();
+        assertThat(cookie.getDomain(), CoreMatchers.equalTo("domain.com"));
+    }
+    
+    @Test
+    public void testThatCookieClearWorks() {
+        String applicationCookieName = ninjaProperties.getOrDie(
+                NinjaConstant.applicationCookiePrefix)
+                + ninja.utils.NinjaConstant.SESSION_SUFFIX;
+        
+        // First roundtrip
+        Session sessionCookie = createNewSession();
+        sessionCookie.init(context);
+        sessionCookie.put("anykey", "anyvalue");
+
+        Session sessionCookieWithValues = roundTrip(sessionCookie);
+        
+        // Second roundtrip with cleared session
+        sessionCookieWithValues.clear();
+        when(context.hasCookie(applicationCookieName)).thenReturn(true);
+        
+        // Third roundtrip
+        String cookieValue = captureFinalCookie(sessionCookieWithValues);
+        assertThat(cookieValue, not(containsString("anykey")));
+        
+        assertThat(cookieCaptor.getValue().getDomain(), CoreMatchers.equalTo(null));
+        assertThat(cookieCaptor.getValue().getMaxAge(), CoreMatchers.equalTo(0));
+    }
+    
+    @Test
+    public void testThatCookieClearWorksWithApplicationDomain() {
+       String applicationCookieName = ninjaProperties.getOrDie(
+                NinjaConstant.applicationCookiePrefix)
+                + ninja.utils.NinjaConstant.SESSION_SUFFIX;
+       when(ninjaProperties.get(NinjaConstant.applicationCookieDomain)).thenReturn("domain.com");
+        
+        // First roundtrip
+        Session sessionCookie = createNewSession();
+        sessionCookie.init(context);
+        sessionCookie.put("anykey", "anyvalue");
+
+        Session sessionCookieWithValues = roundTrip(sessionCookie);
+        
+        // Second roundtrip with cleared session
+        sessionCookieWithValues.clear();
+        when(context.hasCookie(applicationCookieName)).thenReturn(true);
+        
+        // Third roundtrip
+        String cookieValue = captureFinalCookie(sessionCookieWithValues);
+        assertThat(cookieValue, not(containsString("anykey")));
+        
+        assertThat(cookieCaptor.getValue().getDomain(), CoreMatchers.equalTo("domain.com"));
+        assertThat(cookieCaptor.getValue().getMaxAge(), CoreMatchers.equalTo(0));
+    }    
 
     private Session roundTrip(Session sessionCookie1) {
         sessionCookie1.save(context);
@@ -465,7 +545,7 @@ public class SessionImplTest {
         sessionCookie.save(context);
 
         // SessionImpl should set the cookie
-        verify(context).addCookie(cookieCaptor.capture());
+        verify(context, atLeastOnce()).addCookie(cookieCaptor.capture());
 
         String cookieValue = cookieCaptor.getValue().getValue();
         String cookieValueWithoutSign = cookieValue.substring(cookieValue.indexOf("-") + 1);